pySim-shell: add export option to add CREATE FILE APDUs
Export currently only exports the contents of the files. When running
the resulting script it is expected that the files already exist. This
is normaly the case. However, when replaying the script on a blank card
the files do not exist yet.

The CREATE FILE parameters are created based on the FCP template that
the card returns when the file is selected. In some cases the returned
FCP template may lacl data that the card needs to create the file
properly. This means that the --create parameter must be used with
caution and possibly manual re-work of the result is necessary.

Change-Id: If529b86062f7c678c9c75f46ce9ca116cea8b47b